First major commit of CodeExtractor-based FatBinDump.cpp
For the record, current sort-of-ish requirements/known limits:
- Try to keep variable declarations (and their associated AllocaInsts)
near the top of the source file
- Use simple (easy to find as constant) malloc/calloc expressions
(or declare at the top of program and malloc way down later so
that the AllocaInst is still in BB 0 but the complex logic happens inside one of the nodes)
- Structs/array constants/etc support tbd
- No command line arguments (the runtime has no way to ask for them/pass
them in yet anyway)
- Can't have any fancy function calls that initialize values happen in
your first basic block. BB 0 isn't included as a DAG node, so if
something in BB 0 can't be found to be a constant value, it won't be
initialized properly. Just put a dummy for loop and push complex
initializers down into a different BB.