First pass loop-based fixup in KernelWrapper.cpp
This loop-based fixup uses the dominator tree analysis and loop info
analysis to identify loops inside main and then ensure that no loop is
being split across multiple groups of blocks in the baseline program

This was guided by an issue with a kernel being detected as an inner for
loop in a double for loop. The first part of the parent loop was in one
node, split by the kernel node, and the second half was in another node

This second half was unable to be outlined as it required returning to
the loop header which would have meant unconditionally jumping to a BB
contained inside a different function, which the code extractor doesn't
support

As such, find all latches for the loops present in main, and we widen
their block groups to ensure that the entirety of that loop is
encompassed inside the outlined block. This has a number of
disadvantages in terms of optimization opportunities it presents and, in
some cases, makes the kernel-based analysis useless, but it's a path
toward functional code for the time being

In smaller changes, we now only try to outline blocks contained in the main
function rather than all functions in the source LLVM file, and there
are some try-catch blocks added to cartographer/etc to handle
exceptions that are occasionally thrown by std::vector when a trace has
some corruption (see i.e. ruhrie/TraceAtlas #16)